Set Big Button margin to 0 auto
Now I know you can enter two values in one property.
What happens hear is the top and bottom margins are
set to zero, while "auto" is the left and right.
Auto makes it so that it is automatically centered.
I forgot about "auto" so I manually centered the
button by using 120px of margin-left.